+++ xous-rs/src/definitions/messages/envelope.rs
@@ -40,16 +40,6 @@ impl Drop for Envelope {
                     .expect("couldn't return memory")
             }
             Message::Move(msg) => {
-                // Reconstitute the vec and drop it right away to ensure it's freed.
-                #[cfg(any(unix, windows))]
-                let _v = unsafe {
-                    std::vec::Vec::from_raw_parts(
-                        msg.buf.as_mut_ptr(),
-                        msg.buf.len(),
-                        msg.buf.len(),
-                    )
-                };
-                #[cfg(not(any(unix, windows)))]
                 crate::syscall::unmap_memory(msg.buf).expect("couldn't free memory message")
             }
             _ => (),
